+++ src/cli.ts
@@ -2,13 +2,13 @@
 
 import { romanize } from './index';
 
-const commandArg = process.argv[2];
+const commandArg = process.argv[2] as unknown;
 
 if (!commandArg) {
     throw new Error('A number must be given as command parameter.');
 }
 
-const result = romanize(Number(commandArg));
+const result = romanize(commandArg as number);
 
 // tslint:disable-next-line:no-console
 console.log(result);