+++ CHANGELOG.md
@@ -4,10 +4,16 @@ All notable changes to the romanize project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-<!--## [Unreleased]-->
+## [Unreleased]
+### Added
+- Tests for CLI interface.
+
+### Fixed
+- CLI will now output a better error message when input is not a number.
+
 ## [1.0.0] - 2019-04-10
 ### Added
-- Setup for this project
+- Setup for this project.
 
 [Unreleased]: https://github.com/Ionaru/romanize/compare/1.0.0...HEAD 
 [1.0.0]: https://github.com/Ionaru/romanize/compare/fdf1fc5...1.0.0