+++ CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to the romanize project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+<!--## [Unreleased]-->
+## [1.1.0] - 2019-04-11
 ### Added
 - Tests for CLI interface.
 - Docstrings in dist modules.
@@ -17,5 +18,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Setup for this project.
 
-[Unreleased]: https://github.com/Ionaru/romanize/compare/1.0.0...HEAD 
+[Unreleased]: https://github.com/Ionaru/romanize/compare/1.1.0...HEAD 
+[1.1.0]: https://github.com/Ionaru/romanize/compare/1.0.0...1.1.0 
 [1.0.0]: https://github.com/Ionaru/romanize/compare/fdf1fc5...1.0.0